Moved SideNavbar files to new folder
